+++ hal/inc/core_hal.h
@@ -157,7 +157,7 @@ uint32_t HAL_Core_Runtime_Info(runtime_info_t* info, void* reserved);
 
 extern void app_setup_and_loop(void);
 extern void app_setup_and_loop_initial(void);
-extern void app_loop(void);
+extern void app_loop(bool threaded);
 
 typedef enum HAL_SystemClock
 {